requirements.txt
cython==3.0.11
setuptools==67.6.1
python-dotenv==1.0.0
tiktoken==0.8.0
transformers
ipython==8.28.0
torch==2.3.1
openai==1.51.2
wheel==0.44.0
PyAudio==0.2.14
RealtimeSTT==0.3.0  
webrtcvad==2.0.10
numpy==1.26.4
requests==2.32.3
websocket-client==1.8.0
transitions==0.9.2
pygame==2.6.1
google-generativeai==0.8.3
google-cloud-texttospeech==2.21.0
fastapi==0.115.4
uvicorn==0.32.0
websockets==12.0
websocket-client==1.8.0
fastapi-sessions==0.3.2
Jinja2==3.1.4
python-multipart==0.0.17
SpeechRecognition==3.11.0
